requirements.txt
Flask
requests
python-dotenv
gunicorn
Flask-Limiter